DISCO-2147 Add resource validation to add persona page
