+++ angular/src/app/admin/personas/persona-detail/persona-detail.component.ts
@@ -1,6 +1,5 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
-import _get from 'lodash.get';
 
 import { ConfigModificationObject } from '../../shared/configModificationObject';
 import { EntityDetailBase } from '../../shared/entity-detail.base';
@@ -31,7 +30,7 @@ export class PersonaDetailComponent extends EntityDetailBase<PersonaService> imp
     const personaModel: EntityModel = this.personaForm.getEntityModel();
     const change = new ConfigModificationObject(personaModel.dto, {});
     this.entityService.update(this.entity.name, change)
-      .subscribe(this.navigateUp, (err) => this.invalidateAccessFields(err));
+      .subscribe(this.navigateUp, (err) => this.personaForm.invalidateAccessFields(err));
   }
 
   delete() {
@@ -41,23 +40,5 @@ export class PersonaDetailComponent extends EntityDetailBase<PersonaService> imp
 
   private navigateUp = () => this.router.navigate(['..'], { relativeTo: this.route });
 
-  private invalidateAccessFields({error}) {
-    const personaId = this.personaForm.form.get('id').value;
-    const resourcesFormGroup = this.personaForm.form.get('resources');
-    const fieldsToAdd: object = _get(error, `testPersonas[${personaId}].addResources`);
-    const fieldsToRemove: object = _get(error, `testPersonas[${personaId}].removeResources`);
 
-    const setError = ([resourceName, { access }]) => {
-      access.forEach((accessRole) => {
-        const accessRoleCheckbox = resourcesFormGroup.get(resourceName).get(accessRole);
-        accessRoleCheckbox.setErrors({'Doesn\'t match role criteria': true});
-      });
-    };
-
-    Object.entries(fieldsToAdd)
-      .forEach(setError);
-
-    Object.entries(fieldsToRemove)
-      .forEach(setError);
-  }
 }
